передаллал много чего Сделан буферный файл для скоадвания в него ссылок сделан генератор который обрабатывает эт файл
Надо обязательно сделать быстрый способ исключения дубликатов при добавлении в БД.

+++ url_graber.py
@@ -1,4 +1,3 @@
-from bs4 import BeautifulSoup
 import crawl_parser
 import downloader
 import os
@@ -10,16 +9,22 @@ class SitemapGruber():
         self.parser_robots = crawl_parser.RobotsParser()
         self.downloader = downloader.Downloader()
 
-    def get_url_from_sitemap(self, site_url):
+    def get_url_from_sitemap(self, site_url, tmp_file, site_id):
         sitemap_list = self.parser_robots.get_sitemap_list(site_url)
-        url_list = []
-        for sitemap in sitemap_list:
-            file = self.downloader.download_file(sitemap)
-            url_dict = self.parser_sitemap.parse_sitemap(file)
-            sitemap_list.extend(url_dict['sitemap'])
-            url_list.extend(url_dict['urls'])
-            os.remove(file)
-        return url_list
+        if sitemap_list:
+            # url_list = []
+            for sitemap in sitemap_list:
+                file = self.downloader.download_file(sitemap)
+                if file:
+                    url_dict = self.parser_sitemap.parse_sitemap(file, tmp_file, site_id)
+                    sitemap_list.extend(url_dict['sitemap'])
+                    # url_list.extend(url_dict['urls'])
+                    os.remove(file)
+                else:
+                    pass
+            # return url_list
+        else:
+            pass
 
 
 class SiteGruber():
